added implementation to navigate to the forecast page when a location is selected in the recycler view

+++ app/src/main/java/com/pixelart/geocodingweatherapp/data/dto/weather/Main.kt
@@ -16,5 +16,5 @@ data class Main(
     val grndLevel: Double,
     val humidity: Int,
     @SerializedName("temp_kf")
-    val tempKf: Int
+    val tempKf: Double
 )
